Relocate unit test code to match code under test.
